Add test to makefile to prevent TLH in posts

+++ Makefile
@@ -1,4 +1,6 @@
 deploy:
+	# Verify there are no top-level headings in blog posts
+	grep -r '^# ' content/posts/ && exit 1 || true
 	hugo
 	rsync -avx --delete public/ cocinero-tarneo:./www/
